+++ src/Plugins/Module/Location/src/DateItemDelegate.h
@@ -22,20 +22,31 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  */
-#include <QPlainTextEdit>
+#ifndef DATEITEMDELEGATE_H
+#define DATEITEMDELEGATE_H
 
-#include "FocusPlainTextEdit.h"
+#include <QStyledItemDelegate>
+#include <QModelIndex>
 
-// PUBLIC
+class QWidget;
 
-FocusPlainTextEdit::FocusPlainTextEdit(QWidget *parent) noexcept
-    : QPlainTextEdit {parent}
-{}
-
-// PROTECTED
-
-void FocusPlainTextEdit::focusOutEvent(QFocusEvent *event) noexcept
+class DateItemDelegate : public QStyledItemDelegate
 {
-    QPlainTextEdit::focusOutEvent(event);
-    emit focusLost();
-}
+    Q_OBJECT
+public:
+    DateItemDelegate() noexcept;
+    DateItemDelegate(const DateItemDelegate &rhs) = delete;
+    DateItemDelegate(DateItemDelegate &&rhs) = delete;
+    DateItemDelegate &operator=(const DateItemDelegate &rhs) = delete;
+    DateItemDelegate &operator=(DateItemDelegate &&rhs) = delete;
+    ~DateItemDelegate() override;
+
+    QWidget *createEditor(QWidget *parent, const QStyleOptionViewItem &option, const QModelIndex &index) const override;
+    void setEditorData(QWidget *editor, const QModelIndex &index) const noexcept override;
+    void setModelData(QWidget *editor, QAbstractItemModel *model, const QModelIndex &index) const noexcept override;
+
+private slots:
+    void commitAndCloseEditor() noexcept;
+};
+
+#endif // DATEITEMDELEGATE_H